fix: make relative positions absolute
